BR-315 Remove cast that was rounding the special's power number (found by Tyrion)

+++ Quantum/quantum_code/quantum.code/Components/SpecialHazardAimSpawn.cs
@@ -49,7 +49,7 @@ namespace Quantum
 				GameId = special.SpecialId,
 				Interval = special.Speed,
 				NextTickTime = f.Time + special.Speed,
-				PowerAmount = (uint)special.SpecialPower,
+				PowerAmount = special.SpecialPower,
 				Radius = special.Radius,
 				StunDuration = FP._0,
 				TeamSource = team,
